requirements.txt
streamlit
pymupdf
pandas
openpyxl
xlsxwriter
google-generativeai
watchdog